Refine vim installation.

+++ setup.sh
@@ -11,23 +11,41 @@ sudo apt-get install xclip
 echo "Installing ack..."
 sudo apt-get install ack-grep
 
+echo "----------------------------------------------------------"
 echo "Installing vim..."
 # https://github.com/Valloric/YouCompleteMe/wiki/Building-Vim-from-source
-sudo apt-get remove vim vim-runtime gvim vim-tiny vim-common
 mkdir -p ~/software
 cd ~/software
-git clone https://github.com/vim/vim.git
-cd vim
-./configure --with-features=huge --enable-multibyte --enable-rubyinterp --enable-pythoninterp --with-python-config-dir=/usr/lib/python2.7/config --enable-perlinterp --enable-luainterp --enable-gui=gtk2 --enable-cscope --prefix=/usr
-make VIMRUNTIMEDIR=/usr/share/vim/vim74
-sudo make install
-sudo update-alternatives --install /usr/bin/editor editor /usr/bin/vim 1
-sudo update-alternatives --set editor /usr/bin/vim
-sudo update-alternatives --install /usr/bin/vi vi /usr/bin/vim 1
-sudo update-alternatives --set vi /usr/bin/vim
-echo
-vim --help | head -n1
-echo
+if [[ ! -e vim ]]; then
+    sudo apt-get remove -y vim vim-runtime gvim vim-tiny vim-common
+    sudo apt-get install -y libncurses5-dev libgnome2-dev libgnomeui-dev \
+        libgtk2.0-dev libatk1.0-dev libbonoboui2-dev \
+        libcairo2-dev libx11-dev libxpm-dev libxt-dev python-dev \
+        python3-dev
+    git clone https://github.com/vim/vim.git
+    cd vim
+    # TODO: specify correct python config-dir!
+    ./configure --with-features=huge \
+        --enable-multibyte \
+        --enable-pythoninterp \
+        --with-python-config-dir=/usr/lib/python2.7/config-x86_64-linux-gnu \
+        --enable-python3interp \
+        --with-python3-config-dir=/usr/lib/python3.5/config-3.5m-x86_64-linux-gnu \
+        --enable-gui=gtk2 \
+        --enable-cscope \
+        --prefix=/usr
+    # obtain version info, strip quotes (http://stackoverflow.com/questions/9733338/shell-script-remove-first-and-last-quote-from-a-variable)
+    vim_version=`ack '#define VIM_VERSION_NODOT' src/version.h | awk '{ print $3; }' | tr -d '"'`
+    echo $vim_version
+    make VIMRUNTIMEDIR=/usr/share/vim/$vim_version
+    sudo make install
+    sudo update-alternatives --install /usr/bin/editor editor /usr/bin/vim 1
+    sudo update-alternatives --set editor /usr/bin/vim
+    sudo update-alternatives --install /usr/bin/vi vi /usr/bin/vim 1
+    sudo update-alternatives --set vi /usr/bin/vim
+    echo
+    vim --help | head -n1
+fi
 
 echo "Setting up .files..."
 cd ~
